requirements.txt
fusepy==3.0.1
grpcio-tools
pytest
icecream
click
redis
